requirements.txt
scapy-python3
aiohttp
aiohttp_jinja2
